+++ CHANGES.md
@@ -3,6 +3,7 @@
 * [Issue #140](https://github.com/pathikrit/better-files/issues/140): Batch up events for file monitoring
 * [Issue #136](https://github.com/pathikrit/better-files/issues/136): Use execution contexts for file monitoring
 * [Issue #152](https://github.com/pathikrit/better-files/issues/152): Streamed unzipping
+* [Issue #150](https://github.com/pathikrit/better-files/issues/150): `ManagedResource[File]` for temp files
 
 ## v3.0.0